vanity url clean up

+++ client/src/routes/home/VanityURL.jsx
@@ -9,8 +9,6 @@ const queryString = window.location.search;
 const urlParams = new URLSearchParams(queryString);
 const urlParamsKeys = urlParams.keys()
 
-console.log('VanityURL',{currentPath})
-
 export const VanityURL = ({
   record, // page information,
   name // subdomain name
@@ -25,24 +23,22 @@ export const VanityURL = ({
     pageAddress.toLowerCase() === address.toLowerCase()
 
   useEffect(() => {
-    console.log('VanityURL',{name})
     if (!name) {
       return
     }
 
     const call = async () => {
       const redirectURL = await api.getURL(name, currentPath)
-      console.log('VanityURL',{redirectURL})
       if (redirectURL) {
         window.location.href = redirectURL
       }
     }
 
     call()
-  }, [name, connector])
+  }, [name])
 
   useEffect(() => {
-    if (!isOwner) {
+    if (!isOwner || !isConnected) {
       return
     }
 
@@ -51,7 +47,7 @@ export const VanityURL = ({
         const value = urlParams.getAll(key)[0]
         const currentAlias = await api.checkURLValidity(name, key)
 
-        if (!value) {
+        if (!value && currentAlias) {
           // remove alias
           try {
             const newAlias = await api.deleteURL(connector, address, name, key)
@@ -84,7 +80,7 @@ export const VanityURL = ({
     }
 
     call()
-  }, [isOwner])
+  }, [isOwner, isConnected])
 
   return <></>
 }
